refactor(telemetry): Update InMemoryTelemetryBuffer to use TelemetryBufferConfig
- Replaced references to TelemetryBufferPolicy with TelemetryBufferConfig in InMemoryTelemetryBuffer and its tests, ensuring consistent naming and improved clarity.
- Adjusted the constructor and internal logic to utilize the new configuration class for buffer management parameters.
