+++ packages/dart/lib/src/telemetry_processing/telemetry_buffer_config.dart
